# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Patient Workflow System API
  version: 1.1.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /doctors/{id}:
    get: { summary: Get doctor by id, parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { "200": { description: OK }, "404": { description: Not Found } } }
    put: { summary: Update doctor, parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { "200": { description: OK }, "403": { description: Forbidden }, "404": { description: Not Found } } }
    delete: { summary: Delete doctor, parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { "204": { description: No Content }, "404": { description: Not Found } } }
...
  /schedules/{id}:
    get: { summary: Get schedule, parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { "200": { description: OK }, "404": { description: Not Found } } }
    put: { summary: Update schedule, parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { "200": { description: OK } } }
    delete: { summary: Delete schedule, parameters: [{ in: path, name: id, required: true, schema: { type: string } }], responses: { "204": { description: No Content } } }
...